requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==3.0.10
Flask-SocketIO==5.3.6
python-socketio==5.11.0
eventlet==0.33.3
psycopg2-binary==2.9.9
confluent-kafka==2.3.0
python-dotenv==1.0.0
marshmallow==3.20.1
requests==2.31.0
pymongo==4.6.1